# docs/swagger.yaml
basePath: /
definitions:
...
  models.Group:
    properties:
      comment:
        type: string
      description:
        description: Alias for Comment if needed
        type: string
      name:
        type: string
    type: object
...
  models.User:
    properties:
      comment:
        type: string
      disabled:
        type: boolean
      email:
        items:
          type: string
        type: array
      groups:
        description: Groups is a computed field (memberships), useful for API
        items:
          type: string
        type: array
      name:
        description: Display Name / Full Name
        type: string
      password:
        description: |-
          Password is write-only, used for creation/update; credentials are
          managed by the identity backend and never stored by this server.
        type: string
      uid:
        type: integer
      username:
        description: ID / Login
        type: string
    type: object
...
host: localhost:8093
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Minimal API server for OKDP UI New
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: OKDP Server New API
  version: "1.0"
paths:
...
  /api/v1/identity/groups:
    get:
      consumes:
      - application/json
      description: Get all groups from the identity provider (Keycloak)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Group'
            type: array
      summary: List all groups
      tags:
      - identity
    post:
      consumes:
      - application/json
      description: Create a new group in the identity provider (Keycloak)
      parameters:
      - description: Group
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/models.Group'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Group'
      summary: Create a new group
      tags:
      - identity
  /api/v1/identity/groups/{name}:
    delete:
      consumes:
      - application/json
      description: Delete a group from the identity provider (Keycloak)
      parameters:
      - description: Group Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete a group
      tags:
      - identity
    put:
      consumes:
      - application/json
      description: Update a group in the identity provider (Keycloak)
      parameters:
      - description: Group Name
        in: path
        name: name
        required: true
        type: string
      - description: Group
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/models.Group'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Group'
      summary: Update a group
      tags:
      - identity
  /api/v1/identity/users:
    get:
      consumes:
      - application/json
      description: Get all users from the identity provider (Keycloak)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: List all users
      tags:
      - identity
    post:
      consumes:
      - application/json
      description: Create a new user in the identity provider (Keycloak)
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
      summary: Create a new user
      tags:
      - identity
  /api/v1/identity/users/{name}:
    delete:
      consumes:
      - application/json
      description: Delete a user from the identity provider (Keycloak)
      parameters:
      - description: User Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete a user
      tags:
      - identity
    get:
      consumes:
      - application/json
      description: Get a user details
      parameters:
      - description: User Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Get user by name
      tags:
      - identity
    put:
      consumes:
      - application/json
      description: Update a user in the identity provider (Keycloak)
      parameters:
      - description: User Name
        in: path
        name: name
        required: true
        type: string
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Update a user
      tags:
      - identity
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: |-
      OIDC access token, as "Bearer <token>". Required on
      every /api route except /api/capabilities.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"